+++ stripe/params/v2/money_management/_outbound_payment_create_params.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 # File generated from our OpenAPI spec
-from stripe.v2._amount import AmountParam
 from typing import Dict
 from typing_extensions import Literal, NotRequired, TypedDict
 
@@ -11,7 +10,7 @@ _OutboundPaymentCreateParamsBase = TypedDict(
 
 
 class OutboundPaymentCreateParams(_OutboundPaymentCreateParamsBase):
-    amount: AmountParam
+    amount: "OutboundPaymentCreateParamsAmount"
     """
     The "presentment amount" to be sent to the recipient.
     """
@@ -40,6 +39,17 @@ class OutboundPaymentCreateParams(_OutboundPaymentCreateParamsBase):
     to: "OutboundPaymentCreateParamsTo"
     """
     To which payout method to send the OutboundPayment.
+    """
+
+
+class OutboundPaymentCreateParamsAmount(TypedDict):
+    value: NotRequired[int]
+    """
+    A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+    """
+    currency: NotRequired[str]
+    """
+    Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
     """