+++ stripe/params/v2/money_management/_outbound_payment_quote_create_params.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 # File generated from our OpenAPI spec
-from stripe.v2._amount import AmountParam
 from typing_extensions import Literal, NotRequired, TypedDict
 
 _OutboundPaymentQuoteCreateParamsBase = TypedDict(
@@ -10,7 +9,7 @@ _OutboundPaymentQuoteCreateParamsBase = TypedDict(
 
 
 class OutboundPaymentQuoteCreateParams(_OutboundPaymentQuoteCreateParamsBase):
-    amount: AmountParam
+    amount: "OutboundPaymentQuoteCreateParamsAmount"
     """
     The "presentment amount" to be sent to the recipient.
     """
@@ -23,6 +22,17 @@ class OutboundPaymentQuoteCreateParams(_OutboundPaymentQuoteCreateParamsBase):
     to: "OutboundPaymentQuoteCreateParamsTo"
     """
     Request details about the recipient of an OutboundPaymentQuote.
+    """
+
+
+class OutboundPaymentQuoteCreateParamsAmount(TypedDict):
+    value: NotRequired[int]
+    """
+    A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+    """
+    currency: NotRequired[str]
+    """
+    Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
     """