+++ stripe/v2/core/_account.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # File generated from our OpenAPI spec
 from stripe._stripe_object import StripeObject
-from stripe.v2._amount import Amount
 from typing import ClassVar, Dict, List, Optional
 from typing_extensions import Literal
 
@@ -3811,6 +3810,16 @@ class Account(StripeObject):
                 """
 
             class AnnualRevenue(StripeObject):
+                class Amount(StripeObject):
+                    currency: Optional[str]
+                    """
+                    Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+                    """
+                    value: Optional[int]
+                    """
+                    A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+                    """
+
                 amount: Optional[Amount]
                 """
                 A non-negative integer representing the amount in the smallest currency unit.
@@ -3819,6 +3828,7 @@ class Account(StripeObject):
                 """
                 The close-out date of the preceding fiscal year in ISO 8601 format. E.g. 2023-12-31 for the 31st of December, 2023.
                 """
+                _inner_class_types = {"amount": Amount}
 
             class Documents(StripeObject):
                 class BankAccountOwnershipVerification(StripeObject):
@@ -4078,10 +4088,21 @@ class Account(StripeObject):
                 """
 
             class MonthlyEstimatedRevenue(StripeObject):
+                class Amount(StripeObject):
+                    currency: Optional[str]
+                    """
+                    Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+                    """
+                    value: Optional[int]
+                    """
+                    A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+                    """
+
                 amount: Optional[Amount]
                 """
                 A non-negative integer representing the amount in the smallest currency unit.
                 """
+                _inner_class_types = {"amount": Amount}
 
             class ScriptAddresses(StripeObject):
                 class Kana(StripeObject):