+++ stripe/v2/money_management/_outbound_payment_quote.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # File generated from our OpenAPI spec
 from stripe._stripe_object import StripeObject
-from stripe.v2._amount import Amount
 from typing import ClassVar, Dict, List, Optional
 from typing_extensions import Literal
 
@@ -15,6 +14,16 @@ class OutboundPaymentQuote(StripeObject):
         Literal["v2.money_management.outbound_payment_quote"]
     ] = "v2.money_management.outbound_payment_quote"
 
+    class Amount(StripeObject):
+        currency: Optional[str]
+        """
+        Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+        """
+        value: Optional[int]
+        """
+        A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+        """
+
     class DeliveryOptions(StripeObject):
         bank_account: Optional[Literal["automatic", "local", "wire"]]
         """
@@ -22,6 +31,16 @@ class OutboundPaymentQuote(StripeObject):
         """
 
     class EstimatedFee(StripeObject):
+        class Amount(StripeObject):
+            currency: Optional[str]
+            """
+            Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+            """
+            value: Optional[int]
+            """
+            A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+            """
+
         amount: Amount
         """
         The fee amount for corresponding fee type.
@@ -36,9 +55,20 @@ class OutboundPaymentQuote(StripeObject):
         """
         The fee type.
         """
+        _inner_class_types = {"amount": Amount}
 
     class From(StripeObject):
-        debited: Amount
+        class Debited(StripeObject):
+            currency: Optional[str]
+            """
+            Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+            """
+            value: Optional[int]
+            """
+            A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+            """
+
+        debited: Debited
         """
         The monetary amount debited from the sender, only set on responses.
         """
@@ -46,6 +76,7 @@ class OutboundPaymentQuote(StripeObject):
         """
         The FinancialAccount that funds were pulled from.
         """
+        _inner_class_types = {"debited": Debited}
 
     class FxQuote(StripeObject):
         class Rates(StripeObject):
@@ -78,7 +109,17 @@ class OutboundPaymentQuote(StripeObject):
         _inner_class_dicts = ["rates"]
 
     class To(StripeObject):
-        credited: Amount
+        class Credited(StripeObject):
+            currency: Optional[str]
+            """
+            Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+            """
+            value: Optional[int]
+            """
+            A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+            """
+
+        credited: Credited
         """
         The monetary amount being credited to the destination.
         """
@@ -90,6 +131,7 @@ class OutboundPaymentQuote(StripeObject):
         """
         To which account the OutboundPayment is sent.
         """
+        _inner_class_types = {"credited": Credited}
 
     amount: Amount
     """
@@ -133,6 +175,7 @@ class OutboundPaymentQuote(StripeObject):
     Details about the recipient of an OutboundPaymentQuote.
     """
     _inner_class_types = {
+        "amount": Amount,
         "delivery_options": DeliveryOptions,
         "estimated_fees": EstimatedFee,
         "from": From,