+++ stripe/v2/money_management/_transaction_entry.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # File generated from our OpenAPI spec
 from stripe._stripe_object import StripeObject
-from stripe.v2._amount import Amount
 from typing import ClassVar, Optional
 from typing_extensions import Literal
 
@@ -16,18 +15,53 @@ class TransactionEntry(StripeObject):
     )
 
     class BalanceImpact(StripeObject):
-        available: Amount
+        class Available(StripeObject):
+            currency: Optional[str]
+            """
+            Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+            """
+            value: Optional[int]
+            """
+            A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+            """
+
+        class InboundPending(StripeObject):
+            currency: Optional[str]
+            """
+            Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+            """
+            value: Optional[int]
+            """
+            A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+            """
+
+        class OutboundPending(StripeObject):
+            currency: Optional[str]
+            """
+            Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
+            """
+            value: Optional[int]
+            """
+            A non-negative integer representing how much to charge in the [smallest currency unit](https://docs.stripe.com/currencies#minor-units).
+            """
+
+        available: Available
         """
         Impact to the available balance.
         """
-        inbound_pending: Amount
+        inbound_pending: InboundPending
         """
         Impact to the inbound_pending balance.
         """
-        outbound_pending: Amount
+        outbound_pending: OutboundPending
         """
         Impact to the outbound_pending balance.
         """
+        _inner_class_types = {
+            "available": Available,
+            "inbound_pending": InboundPending,
+            "outbound_pending": OutboundPending,
+        }
 
     class TransactionDetails(StripeObject):
         class Flow(StripeObject):